+++ apps/examples/package.json
@@ -34,7 +34,6 @@
 	},
 	"dependencies": {
 		"@playwright/test": "^1.38.1",
-		"@radix-ui/react-accordion": "^1.1.2",
 		"@radix-ui/react-alert-dialog": "^1.0.5",
 		"@tldraw/assets": "workspace:*",
 		"@tldraw/tldraw": "workspace:*",
