+++ apps/examples/src/examples/basic/README.md
@@ -1,0 +1,12 @@
+---
+title: Basic
+component: ./BasicExample.tsx
+category: basic
+priority: 1
+---
+
+...
+
+---
+
+The `Tldraw` component is a regular React component. If the `persistenceKey` is provided, then the component will persist its local state under that key.
