--- apps/examples/src/examples/develop/README.md
@@ -1,12 +1,0 @@
----
-title: Basic
-component: ./BasicExample.tsx
-category: basic
-priority: 1
----
-
-The easiest way to get started with tldraw.
-
----
-
-The simplest use of the `<Tldraw/>` component.
