+++ apps/examples/src/examples/yjs/README.md
@@ -4,9 +4,3 @@ component: ./YjsExample.tsx
 category: collaboration
 priority: 2
 ---
-
-TODOTODO
-
----
-
-TODOTODO
